Conflicts:
	src/main/java/hivemall/knn/similarity/CosineSimilarityUDF.java
	src/test/java/hivemall/knn/similarity/CosineSimilarityUDFTest.java

+++ src/main/java/hivemall/io/SynchronizedModelWrapper.java
@@ -53,7 +53,7 @@ public final class SynchronizedModelWrapper implements PredictionModel {
     }
 
     @Override
-    public int getNumMixed() {
+    public long getNumMixed() {
         return model.getNumMixed();
     }
